plugins/bibliogon-plugin-export/pyproject.toml
[tool.poetry]
name = "bibliogon-plugin-export"
version = "1.0.0"
description = "EPUB, PDF and project structure export plugin for Bibliogon"
authors = ["astrapi69"]
license = "MIT"
readme = "README.md"
packages = [{include = "bibliogon_export"}]

[tool.poetry.dependencies]
python = "^3.11"
pluginforge = {path = "../../pluginforge", develop = true}
fastapi = "^0.115.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[tool.poetry.plugins."bibliogon.plugins"]
export = "bibliogon_export.plugin:ExportPlugin"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
